long is not necessarily 32 bits thus use int for int32_t

+++ src/lua/luaconf.h
@@ -878,7 +878,7 @@ union luai_Cast { double l_d; long l_l; };
 
 #if !defined(LUA_CROSS_COMPILER)
 typedef short int16_t;
-typedef long int32_t;
+typedef int int32_t;
 #endif
 
 /* If you define the next macro you'll get the ability to set rotables as
